requirements.txt
streamlit
requests
beautifulsoup4
groq
exa-py
posthog
python-dotenv
